// src/types/quran.ts
export interface QuranVerse {
  id: number
  jozz: number
  sura_no: number
  sura_name_en: string
  sura_name_ar: string
  page: number
  line_start: number
  line_end: number
  aya_no: number
  aya_text: string
  aya_text_emlaey: string
}

export interface Language {
  code: string
  name: string
  direction: "rtl" | "ltr"
}